🛡️ Sentinel: [HIGH] Fix path traversal in runner CWD validation
